+++ src/main/java/me/jcala/jmooc/service/inter/CrsSer.java
@@ -11,9 +11,9 @@ public interface CrsSer {
 
     long addCourse(Course course);//添加课程
 
-    Set<Chapter> getChapterList(int crsId);//获取某课程的章节列表
+    Set<Chapter> getChapterList(long crsId);//获取某课程的章节列表
 
-    Set<Lesson> getLessonList(int chpId);//获取某章节的课时列表
+    Set<Lesson> getLessonList(long chpId);//获取某章节的课时列表
 
     void addChapter(ChpForm chpForm);
 }
